feat: move pages to /app
- /mes-jeunes/[id_jeune]/favoris
- /mes-jeunes/[id_jeune]/indicateurs
- /mes-jeunes/[id_jeune]/actions/[idAction]
