+++ tests/pages/fiche-jeune/Actions.test.tsx
@@ -4,7 +4,7 @@ import { DateTime } from 'luxon'
 import { useRouter } from 'next/navigation'
 import React from 'react'
 
-import FicheBeneficiairePage from 'app/(connected)/(with-sidebar)/(with-chat)/mes-jeunes/[jeune_id]/FicheBeneficiairePage'
+import FicheBeneficiairePage from 'app/(connected)/(with-sidebar)/(with-chat)/mes-jeunes/[idJeune]/FicheBeneficiairePage'
 import {
   desActionsInitiales,
   desCategories,
