+++ tests/pages/fiche-jeune/Situations.test.tsx
@@ -1,7 +1,7 @@
 import { act, screen } from '@testing-library/react'
 import React from 'react'
 
-import FicheBeneficiairePage from 'app/(connected)/(with-sidebar)/(with-chat)/mes-jeunes/[jeune_id]/FicheBeneficiairePage'
+import FicheBeneficiairePage from 'app/(connected)/(with-sidebar)/(with-chat)/mes-jeunes/[idJeune]/FicheBeneficiairePage'
 import { desActionsInitiales } from 'fixtures/action'
 import { unAgenda } from 'fixtures/agenda'
 import { desIndicateursSemaine, unDetailJeune } from 'fixtures/jeune'
